+++ CHANGELOG.rst
@@ -1,6 +1,11 @@
 Changelog
 =========
 
+Development
+-----------
+
+* Delete SQLite database if, upon program exit, it contains no records. (`#14 <https://github.com/anthonydb/pneumatic/issues/14>`_)
+
 0.1.6 - April 28, 2016
 ----------------------